aff/src/requirements.txt
# Huvudbibliotek
streamlit==1.35.0
openai==1.78.1
requests==2.31.0
fpdf==1.7.2
Pillow==10.0.1
matplotlib==3.9.0
numpy==1.26.0
python-dotenv==1.0.0
plotly==6.0.1

# Datamanipulering och analys
pandas==2.2.0

# Kartvisualisering
folium==0.15.0
